Make StaleLettersTask use StaleLetterService
